requirements.txt
pytorch-lightning==0.9.0
wandb==0.10.11
dotmap
pandas
nlpaug
matplotlib
torch_dct
scikit-learn==0.24.1
librosa==0.9.1
protobuf==3.20.*
click 
requests 
tqdm 
pyspng 
ninja 
imageio-ffmpeg==0.4.3
numpy==1.23.4
